Trait leo_core::algorithms::CoreFunction

trait CoreFunction {
    const NUM_ARGS: usize;

    fn return_type() -> Type;

    fn first_arg_is_allowed_type(_: &Type) -> bool { ... }
    fn second_arg_is_allowed_type(_: &Type) -> bool { ... }
}

A core function of a core struct, e.g. hash or commit Provides required type information to the type checker.

Required Associated Constants

const NUM_ARGS: usize

Required Methods

fn return_type() -> Type

The return type of the core function.

Provided Methods

fn first_arg_is_allowed_type(_: &Type) -> bool

Returns whether or not the first argument is an allowed type.

fn second_arg_is_allowed_type(_: &Type) -> bool

Returns whether or not the second argument is an allowed type.