Trait leo_core::algorithms::CoreFunction

trait CoreFunction {
    const NUM_ARGS: usize;

    fn first_arg_types() -> &'static [Type];
    fn return_type() -> Type;

    fn second_arg_types() -> &'static [Type] { ... }
}

A core function of a core circuit, e.g. hash or commit Provides required type information to the type checker.

Required Associated Constants

const NUM_ARGS: usize

Required Methods

fn first_arg_types() -> &'static [Type]

Returns first argument allowed types.

fn return_type() -> Type

The return type of the core function.

Provided Methods

fn second_arg_types() -> &'static [Type]

Returns the second argument allowed types. Implementing this method is optional since some functions may not allow a second argument.